quantlib-wasm-demo: vitejs, refs #3823
