+++ package.json
@@ -4,27 +4,29 @@
     "homepage": "https://captorab.github.io/quantlib-wasm-demo/",
     "private": false,
     "scripts": {
-        "start": "react-scripts --openssl-legacy-provider start",
-        "build": "react-scripts --openssl-legacy-provider build",
-        "test": "react-scripts  --openssl-legacy-provider test",
+        "start": "vite --port 3000",
+        "build": "vite build",
+        "test": "cross-env NODE_PATH=. npx jest --detectOpenHandles --maxWorkers=1 --forceExit --verbose",
         "predeploy": "npm run build",
         "deploy": "gh-pages -d build -r https://github.com/CaptorAB/quantlib-wasm-demo"
     },
     "dependencies": {
-        "@captor/react-scripts": "4.0.1",
         "@testing-library/jest-dom": "^5.12.0",
         "@testing-library/react": "^11.2.6",
         "@testing-library/user-event": "^13.1.8",
+        "@vitejs/plugin-react": "^4.3.1",
         "axios": "^0.21.1",
-        "plotly.js-basic-dist": "^1.58.4",
+        "plotly.js-finance-dist": "^2.34.0",
         "qs": "^6.10.1",
         "quantlib-wasm": "1.29.1",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
         "react-plotly.js": "^2.5.1",
+        "vite": "^5.3.5",
         "web-vitals": "^1.1.2"
     },
     "devDependencies": {
+        "@vitejs/plugin-react-swc": "^3.7.0",
         "cpy-cli": "^3.1.1",
         "eslint-config-prettier": "8.3.0",
         "gh-pages": "^3.1.0"
