--- src/quantlib-wasm-wrapper.js
@@ -1,3 +1,0 @@
-import wasm from "quantlib-wasm/dist/quantlib.wasm";
-
-export { wasm };
