+++ vite.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react-swc";
+
+// https://vitejs.dev/config/
+export default defineConfig(({ mode }) => {
+    return {
+        plugins: [react()],
+        build: {
+            sourcemap: true,
+        },
+    };
+});
